modify proxy route to support graphql
